Remove setting values in contructor. Closes #3
