+++ src/reactiveFunction.js
@@ -2,7 +2,7 @@ const parse = inputsStr => inputsStr
   .split(',')
   .map(input => input.trim());
 
-export default ([fn, inputsStr]) => ({
+export const ReactiveFunction = (fn, inputsStr) => ({
   fn,
   inputs: parse(inputsStr)
 })
